examples/python/requirements.txt
absl-py
google-cloud-core
google-cloud-storage
